Changes : - Organaized if condition @ /employees api - passed token to getEmpImg function to fix wrong "Token Expired Image" - abstracted all perms implementation to perms class to centralized - Created A Transaction Function to execute Multiple SQls and rollback   on error   - Fixed setCachedEmployees @ delete user to filter user from cached     array    - Passed paramsArray to be more secure from SQL INJECTIONS
  Next Goals:
  - Preview Number of Pages to user and let him jump to certain page

+++ Server/Classes/Admin.js
@@ -1,20 +1,14 @@
 const User = require("./User");
-const executeMySqlQuery = require("../Utils/executeMySqlQuery");
-const stringifyFields = require("../Utils/stringifyFields");
 const roles = require("./roles");
+const perms = require("./perms");
 const consoleLog = require("../Utils/consoleLog");
 /*
-
 Admin should be able to do
-
-
 */
 class Admin extends User {
     static priority = 50; // we are going to use priority to check that user is editing people with lower or equal priority
-    constructor(emp_email,role){
-        super(emp_email,role);
-        
-    }
+    constructor(){}
+    
     getPriority(){
         return this.priority
     }
@@ -25,10 +19,8 @@ class Admin extends User {
             try{
                
             if( this.priority >= roles.getRolePriority(otherUserRole)){
-                const fields = stringifyFields("joined",entries);
-                const query = `UPDATE employees SET ${fields} WHERE emp_id = ${emp_id}`
-                await executeMySqlQuery(query ,"Error Updating User Role");
 
+                await perms.executeEditOthers(emp_id  , entries)
                 resolve(true);
             }
             else{
@@ -44,30 +36,23 @@ class Admin extends User {
         
     }
     // other user must be admin or less role, cannot be superAdmin
-    async RemoveOtherUser(emp_id){
+    async RemoveOtherUser(emp_id , otherUserRole){
         if( this.priority >= roles.getRolePriority(otherUserRole)){
 
-        // remove user data & refrenced role & perms
-        const queries = [`DELETE FROM employee_perms WHERE emp_id = ${emp_id}` , `DELETE FROM roles WHERE emp_id = ${emp_id}` , `DELETE FROM employees WHERE emp_id = ${emp_id}` ]
-        const removeOtherUserPromises = [];
-        queries.forEach((query)=>{
-            const promise = new Promise(async (resolve , reject)=>{
-                try{
-                    const result = await executeMySqlQuery(query ,"Error Updating User Role" );
-                    resolve(result);
-                }
-                catch(err){
-                    reject(err);
-                }
-            })
+             /*
+            This order ensures correct deletion to avoid deleting row with refrenced key error
+            ----------------------------------------------------------------------------------
+            Step 1 remove user from employee_perms
+            Step 2 remove user from roles
+            Step 3 remove user from employees
+            */
 
-            removeOtherUserPromises.push(promise);
-            
-        })
+            // this is a quick check of emp_id is number to prevent sql injection
+            if(isNaN(Number(emp_id))) return;
 
-        settled =await Promise.allSettled(removeOtherUserPromises);
-        // to return true if all promises fulfilled
-        return settled.filter((status)=> status !== "rejected").length === queries.length;
+            const queries = [ `DELETE FROM employee_perms WHERE emp_id = ${emp_id}` , `DELETE FROM roles WHERE emp_id = ${emp_id}` , `DELETE FROM employees WHERE emp_id = ${emp_id}` ]
+
+            return await perms.executeRemoveOtherPerm(queries);
 
         }
         else{
